requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
google-generativeai==0.3.2
openai==1.12.0
requests==2.31.0
PyPDF2==3.0.1
pymupdf==1.23.8
markdown2==2.4.12
huggingface-hub==0.20.3
gunicorn==21.2.0
